Implemented methods on TournamentService: delete(), getAll(), update()

+++ src/main/java/com/ila/checkmatecentral/exceptions/TournamentNotFoundException.java
@@ -2,7 +2,7 @@ package com.ila.checkmatecentral.exceptions;
 
 public class TournamentNotFoundException extends RuntimeException{
     
-    public TournamentNotFoundException(Long tournamentId) {
+    public TournamentNotFoundException(Integer tournamentId) {
         super("Could not find Tournament " + tournamentId);
     }
